requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
coveralls==1.8.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==1.11.24
django-bootstrap4==1.0.1
django-chartjs==1.5.0
docopt==0.6.2
eventbrite==3.3.3
gunicorn==19.9.0
idna==2.8
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.3
urllib3==1.25.3
whitenoise==4.1.3
